chore: refactor code locations + names
