+++ src/components/incentives/IncentivesButton.tsx
@@ -48,50 +48,6 @@ export const BlankIncentives = () => {
   );
 };
 
-export const useAllIncentives = ({
-  symbol,
-  rewardedAsset,
-  market,
-  protocolAction,
-  lmIncentives,
-}: {
-  symbol: string;
-  market: string;
-  rewardedAsset?: string;
-  protocolAction?: ProtocolAction;
-  lmIncentives?: ReserveIncentiveResponse[];
-}) => {
-  const { data: meritIncentives } = useMeritIncentives({
-    symbol,
-    market,
-    protocolAction,
-  });
-  const { data: zkSyncIgniteIncentives } = useZkSyncIgniteIncentives({
-    market,
-    rewardedAsset,
-    protocolAction,
-  });
-  const lmIncentivesFiltered = lmIncentives?.filter((i) => i.incentiveAPR !== '0');
-
-  const meritApr =
-    meritIncentives && meritIncentives.incentiveAPR ? Number(meritIncentives?.incentiveAPR) : 0;
-  const zkSyncApr =
-    zkSyncIgniteIncentives && zkSyncIgniteIncentives.incentiveAPR
-      ? Number(zkSyncIgniteIncentives?.incentiveAPR)
-      : 0;
-  const lmApr = lmIncentivesFiltered?.reduce((a, b) => a + +b.incentiveAPR, 0) ?? 0;
-
-  const totalApr = meritApr + zkSyncApr + lmApr;
-
-  const allIncentives = [
-    ...(meritIncentives ? [meritIncentives] : []),
-    ...(zkSyncIgniteIncentives ? [zkSyncIgniteIncentives] : []),
-    ...(lmIncentivesFiltered || []),
-  ];
-
-  return { allIncentives, totalApr };
-};
-
 export const MeritIncentivesButton = (params: {
   symbol: string;
   market: string;
@@ -168,7 +124,7 @@ export const NoAprExternalIncentiveTooltip = ({
   );
 };
 
-export const IncentivesButton = ({ incentives, symbol, displayBlank }: IncentivesButtonProps) => {
+export const LmIncentivesButton = ({ incentives, symbol, displayBlank }: IncentivesButtonProps) => {
   const [open, setOpen] = useState(false);
 
   if (!(incentives && incentives.filter((i) => i.incentiveAPR !== '0').length > 0)) {
